telegram_reader/requirements.txt
telethon>=1.30.3
aiohttp>=3.8.0
python-dotenv>=1.0.0